+++ plone/app/z3cform/csrf.py
@@ -1,9 +1,9 @@
+from plone.app.z3cform.interfaces import IPloneFormLayer
+from plone.protect import CheckAuthenticator
 from z3c.form.button import ButtonActions
 from z3c.form.interfaces import IButtonForm
 from zope.component import adapts
 from zope.interface import Interface
-from plone.app.z3cform.interfaces import IPloneFormLayer
-from plone.protect import CheckAuthenticator
 
 
 class AuthenticatedButtonActions(ButtonActions):